Refactor exceptions
CommandTimeOutError and InitError moved to exceptions.py, DeviceTimeOutError removed.

InitError renamed CommanduinoInitError
CommandTimeOutError renamed CommanduinoCommandTimeOutError

DeviceTimeOutError was defined, imported a couple of times but never used.
