+++ commanduino/commanddevices/commandaccelstepper.py
@@ -9,7 +9,6 @@
 """
 
 from .commanddevice import CommandDevice
-from .commanddevice import DeviceTimeOutError
 
 import time
 import logging
